refactor: destructure GearCard slop props, extract scroll management, and fix image aspect ratios
- Updated `GearCard.tsx` to safely destructure unused resource props out of the `rest` object (`cleanProps`), replacing the brittle `SLOP_PROPS` deletion loop.
- Extracted 'God Effect' scroll management and gesture logic from `MainLayout.tsx` into a new `useScrollManagement.ts` custom hook.
- Fixed CLS and layout shifts by setting explicit `width` and `height` properties on images in `ArielProfile.tsx`, `GearCard.tsx`, and `DetailLayout.tsx` matching their actual intrinsic aspect ratios.
